refactor: ParticipantProfits 클래스 이름 수정, 비즈니스 로직 이동
- ProfitsDto 클래스 이름으로 변경
- record 변경
- 비즈니스 로직 BlackjackGame로 이동
