+++ src/main/java/model/result/ProfitsDto.java
@@ -1,0 +1,7 @@
+package model.result;
+
+import java.util.List;
+
+public record ProfitsDto(ProfitDto dealerProfit, List<ProfitDto> playerProfits) {
+
+}
