+++ src/test/java/model/game/BlackjackGameTest.java
@@ -1,16 +1,18 @@
 package model.game;
 
+import static model.card.CardNumber.ACE;
 import static model.card.CardNumber.JACK;
+import static model.card.CardNumber.KING;
 import static model.card.CardNumber.SEVEN;
 import static model.card.CardNumber.SIX;
 import static model.card.CardNumber.TEN;
 import static model.card.CardNumber.TWO;
 import static model.card.CardShape.CLOVER;
 import static model.card.CardShape.HEART;
+import static model.card.CardShape.SPADE;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
-import java.math.BigDecimal;
 import java.util.List;
 import model.betting.Bet;
 import model.card.Card;
@@ -20,10 +22,10 @@ import model.participant.Player;
 import model.participant.Players;
 import model.result.ParticipantCard;
 import model.result.ParticipantCards;
-import model.result.ProfitDto;
-import model.result.ParticipantProfits;
 import model.result.ParticipantScore;
 import model.result.ParticipantScores;
+import model.result.ProfitDto;
+import model.result.ProfitsDto;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -102,27 +104,33 @@ class BlackjackGameTest {
         }
     }
 
-    @DisplayName("플레이어의 승무패 결과에 따라 딜러의 총 수익 계산")
+    @DisplayName("딜러와 참가자들의 배팅 수익 결과 생성")
     @Test
     void calculateProfit() {
-        Players players = Players.from(List.of("조조"));
-        prepareBets(players);
-        Dealer dealer = new Dealer();
+        Players players = preparePlayers();
+        Dealer dealer = new Dealer(List.of(new Card(KING, HEART), new Card(JACK, SPADE)));
         BlackjackGame blackjackGame = new BlackjackGame();
-        blackjackGame.dealInitialCards(dealer, players);
 
-        ParticipantProfits participantProfits = blackjackGame.calculateProfit(dealer, players);
-        BigDecimal dealerProfit = participantProfits.getPlayerProfits()
-            .stream()
-            .map(ProfitDto::profit)
-            .reduce(BigDecimal.ZERO, BigDecimal::subtract);
+        ProfitsDto profitsDto = blackjackGame.calculateProfit(dealer, players);
 
-        assertThat(dealerProfit).isEqualTo(participantProfits.getDealerProfit().profit());
+        ProfitDto dealerProfit = profitsDto.dealerProfit();
+        ProfitDto jojoProfit = profitsDto.playerProfits().get(0);
+
+        assertAll(
+            () -> assertThat(jojoProfit.profit().intValue()).isEqualTo(15000),
+            () -> assertThat(dealerProfit.profit().intValue()).isEqualTo(-15000)
+        );
     }
 
-    private void prepareBets(Players players) {
+    private Players preparePlayers() {
+        Players players = Players.from(List.of("조조"));
+        Player jojo = new Player("조조", List.of(new Card(ACE, HEART), new Card(JACK, SPADE)));
+        players.getPlayers()
+            .set(0, jojo);
+
         players.getPlayers()
             .get(0)
             .setBet(new Bet("10000"));
+        return players;
     }
 }
